src/backend/pyproject.toml
[project]
name = "integration-copilot-api"
version = "0.1.0"
description = "Integrisight.ai API — a multitenant SaaS backend powered by FastAPI"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]",
    "uvicorn[standard]",
    "pydantic>=2.0",
    "pydantic-settings",
    "aiohttp",
    "azure-identity",
    "azure-cosmos",
    "azure-storage-blob",
    "azure-eventgrid",
    "structlog",
    "opentelemetry-api",
    "opentelemetry-sdk",
    "opentelemetry-instrumentation-fastapi",
    "opentelemetry-instrumentation-logging",
    "opentelemetry-instrumentation-httpx",
    "opentelemetry-instrumentation-aiohttp-client",
    "azure-monitor-opentelemetry",
    "python-jose[cryptography]>=3.5.0",
    "azure-messaging-webpubsubservice>=1.3.0",
    "pyyaml>=6.0.3",
    "python-ulid>=3.0.0",
    "azure-ai-projects>=2.0.0",
]

[dependency-groups]
dev = [
    "ruff",
    "pytest",
    "pytest-asyncio",
    "httpx",
    "pip-audit",
]

[tool.pytest.ini_options]
testpaths = ["../../tests/backend"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
